Move album to common folder
