+++ app/javascript/components/common/album/index.js
