+++ app/javascript/components/discovery/index.js
@@ -1,6 +1,6 @@
 import React, { Fragment, useEffect, useState } from 'react';
 import { Columns, Heading } from 'react-bulma-components';
-import Album from '../album';
+import Album from '../common/album';
 import AlbumsService from '../../services/albums';
 import styled from 'styled-components';
