+++ app/javascript/routes.js
@@ -12,7 +12,7 @@ const Routes = (props) => (
     <Route exact path='/' component={HomeScreen} />
     <Route exact path='/discovery' component={DiscoveryScreen} />
     <Route exact path='/search' component={SearchScreen} />
-    <Route exact path='/album/:id' component={AlbumScreen} />
+    <Route exact path='/albums/:id' component={AlbumScreen} />
     <Route exact path='/favorites' component={FavoritesScreen} />
   </Switch>
 );
